requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
datasets>=2.14.0
trl>=0.7.4
pyyaml>=6.0
huggingface_hub>=0.19.0
sentencepiece>=0.1.99
tokenizers>=0.14.0
safetensors>=0.4.0